chore: add lambdas to expression parser for fees
